requirements.txt
# Dependências do projeto URL Tratamento (Streamlit + Databricks + VTEX)

# --- Core ---
pandas>=2.0.0
chardet>=5.0.0
python-dotenv>=1.0.0

# --- Databricks ---
databricks-sql-connector>=3.0.0

# --- Streamlit (novo frontend) ---
streamlit>=1.32.0
plotly>=5.0.0

# --- HTTP / VTEX ---
requests>=2.31.0
urllib3>=2.0.0

# --- Logging ---
structlog>=24.0.0
